[flp] переделка прогресс-бокса в панелях (теперь не перерисовывается полностью по поводу и без); [localfs] ликвидирована возможность падения при загрузке каталогов; [ve] VE теперь выводит прогресс загрузки файла [FileProcessDialog] возвращена на место рамка окна

+++ pluginner/IFSPlugin.cs
@@ -63,7 +63,8 @@ namespace pluginner
         /// <param name='URL'>
         /// URL of the file.
         /// </param>
-        File GetFile(string URL, int Progress);
+        /// <param name="Progress">OBDOLETE, WILL BE REMOVED SOON</param>
+        File GetFile(string URL, double Progress);
 
         /// <summary>
         /// Gets the file's attribbutes
